+++ src-tauri/src/main.rs
@@ -1,6 +1,11 @@
 // Prevents additional console window on Windows in release, DO NOT REMOVE!!
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
+mod config_data;
+mod error;
+mod nightscout;
+mod utils;
+
 static UPDATE_GLUCOSE_EVENT_ID: &str = "update_glucose";
 static TRAY_MENU_ITEM_GLUCOSE_ENTRY: &str = "tray_menu_item_glucose_entry";
 static TRAY_MENU_ITEM_OPEN_SETTINGS: &str = "tray_menu_item_settings";
@@ -10,104 +15,18 @@ static TRAY_MENU_ITEM_QUIT_DISPLAY: &str = "Quit Glucmon Completely";
 
 use config_data::{get_glucmon_config, set_glucmon_config, GlucmonConfigStore};
 use nightscout::{get_glucose_data, Direction};
-use std::path::PathBuf;
 use std::{sync::Mutex, thread};
 use tauri::{
-    AppHandle, CustomMenuItem, Manager, SystemTray, SystemTrayEvent, SystemTrayMenu,
-    SystemTrayMenuItem, UserAttentionType,
+    CustomMenuItem, Manager, SystemTray, SystemTrayEvent, SystemTrayMenu, SystemTrayMenuItem,
+    UserAttentionType,
 };
+use utils::{create_icon_from_path, get_icon_path_from_direction};
 
 #[derive(Debug)]
 struct Storage {
     config: Mutex<GlucmonConfigStore>,
 }
 
-mod config_data;
-mod nightscout;
-
-use tauri::Icon;
-
-fn create_icon_from_path(path: &PathBuf) -> Result<Icon, Box<dyn std::error::Error>> {
-    let image = image::open(path)?.to_rgba8();
-    let (width, height) = image.dimensions();
-    let resized_image = image::imageops::resize(
-        &image,
-        width / 2,
-        height / 2,
-        image::imageops::FilterType::Nearest,
-    );
-    Ok(Icon::Rgba {
-        rgba: resized_image.into_vec(),
-        width: width / 2,
-        height: height / 2,
-    })
-}
-
-fn get_icon_path_from_direction(
-    app: &AppHandle,
-    direction: &Direction,
-    glucose_value: &str,
-) -> PathBuf {
-    let base_path = "icons/tray/";
-
-    // Parse the glucose_value string to f64
-    let parts: Vec<&str> = glucose_value.split_whitespace().collect();
-    let glucose_str = parts[0];
-    let glucose_f64 = match glucose_str.parse::<f64>() {
-        Ok(value) => value,
-        Err(_) => {
-            return app
-                .path_resolver()
-                .resolve_resource(format!("{}glucmon_icon_NOT-CONFIGURED.png", base_path))
-                .expect("failed to resolve resource")
-        }
-    };
-
-    let severity = match glucose_f64 {
-        v if v < 3.0 => "urgent",
-        v if v < 3.885 => "concern",
-        v if v < 10.0 => "normal",
-        v if v < 13.875 => "concern",
-        _ => "urgent",
-    };
-
-    let direction_str = match direction {
-        Direction::Flat => "flat",
-        Direction::FortyFiveUp => "1_up",     // "fortyfive_up",
-        Direction::FortyFiveDown => "1_down", // "fortyfive_down",
-        Direction::SingleUp => "1_up",
-        Direction::SingleDown => "1_down",
-        Direction::DoubleUp => "2_up",
-        Direction::DoubleDown => "2_down",
-        Direction::TripleUp => "3_up",
-        Direction::TripleDown => "3_down",
-        Direction::RateOutOfRange => {
-            return app
-                .path_resolver()
-                .resolve_resource(format!("{}glucmon_icon_NOT-WORKING.png", base_path))
-                .expect("failed to resolve resource")
-        }
-        Direction::NotComputable => {
-            return app
-                .path_resolver()
-                .resolve_resource(format!("{}glucmon_icon_NOT-WORKING.png", base_path))
-                .expect("failed to resolve resource")
-        }
-        Direction::None => {
-            return app
-                .path_resolver()
-                .resolve_resource(format!("{}glucmon_icon.png", base_path))
-                .expect("failed to resolve resource")
-        }
-    };
-
-    app.path_resolver()
-        .resolve_resource(format!(
-            "{}tray_{}_{}.png",
-            base_path, severity, direction_str
-        ))
-        .expect("failed to resolve resource")
-}
 fn main() {
     let tray_menu_glucose_data_item = CustomMenuItem::new(TRAY_MENU_ITEM_GLUCOSE_ENTRY, "--");
     let tray_menu_settings_item = CustomMenuItem::new(
