+++ src-tauri/src/utils/mod.rs
@@ -1,0 +1,5 @@
+mod create_icon_from_path;
+mod get_icon_path_from_direction;
+
+pub use create_icon_from_path::create_icon_from_path;
+pub use get_icon_path_from_direction::get_icon_path_from_direction;
